jagar: Import bluetooth profile props from stock firmware
Change-Id: Id9c901ed964fedf9cc666e6b2faa94cdadb602ed

jagar: Use the ?= operator so BT properties can be overridden

System properties set with the = operator cannot be overriden. This
device is used elsewhere in the system and some of the properties need
to be overridden. The ?= operator will use the assigned value if no
other values are assigned elsewhere, allowing for overrides.

Bug: 230638462
Test: adb shell getprop | grep bluetooth
Change-Id: Id6cd7901555edafef141bbc1bbd5e06dc3a54d0c